requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.9.0
bidict=0.23.1=pypi_0
blas=2.135=mkl
blas-devel=3.9.0=35_h85df5b5_mkl
blinker=1.9.0=pypi_0
bzip2=1.0.8=h0ad9c76_8
ca-certificates=2025.9.9=haa95532_0
click=8.3.0=pypi_0
colorama=0.4.6=pyhd8ed1ab_1
cpuonly=2.0=0
-e ./python
# dp=0.1=pypi_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
filelock=3.19.1=pyhd8ed1ab_0
flask=3.1.2=pypi_0
flask-socketio=5.5.1=pypi_0
h11=0.16.0=pypi_0
iniconfig=2.0.0=pyhd8ed1ab_1
intel-openmp=2024.2.1=h57928b3_1083
itsdangerous=2.2.0=pypi_0
jinja2=3.1.6=pyhd8ed1ab_0
json5=0.12.1=pypi_0
libblas=3.9.0=35_h5709861_mkl
libcblas=3.9.0=35_h2a3cdd5_mkl
libexpat=2.7.1=hac47afa_0
libffi=3.4.6=h537db12_1
libhwloc=2.12.1=default_h64bd3f2_1002
libiconv=1.18=hc1393d2_2
liblapack=3.9.0=35_hf9ab0e9_mkl
liblapacke=3.9.0=35_h3ae206f_mkl
liblzma=5.8.1=h2466b09_2
liblzma-devel=5.8.1=h2466b09_2
libsqlite=3.50.4=hf5d6505_0
libuv=1.51.0=hfd05255_1
libwinpthread=12.0.0.r4.gg4f2fc60ca=h57928b3_9
libxml2=2.15.0=h5d26750_1
libxml2-16=2.15.0=h692994f_1
libzlib=1.3.1=h2466b09_2
markupsafe=3.0.2=py311h5082efb_1
mkl=2024.2.2=h66d3029_15
mkl-devel=2024.2.2=h57928b3_15
mkl-include=2024.2.2=h66d3029_15
mpmath=1.3.0=pyhd8ed1ab_1
networkx=3.5=pyhe01879c_0
numpy=1.26.0=py311h0b4df5a_0
openssl=3.5.3=h725018a_1
packaging=25.0=pyh29332c3_1
pandas=2.3.3=pypi_0
pip=25.2=pyh8b19718_0
pluggy=1.6.0=pyhd8ed1ab_0
pybind11=2.11.1=py311h005e61a_2
pybind11-global=2.11.1=py311h005e61a_2
pygments=2.19.2=pyhd8ed1ab_0
pytest=8.4.2=pyhd8ed1ab_0
python=3.11.6=h2628c8c_0_cpython
python-dateutil=2.9.0.post0=pypi_0
python-engineio=4.12.3=pypi_0
python-socketio=5.14.1=pypi_0
python_abi=3.11=8_cp311
pytorch=2.1.0=py3.11_cpu_0
pytorch-mutex=1.0=cpu
pytz=2025.2=pypi_0
pyyaml=6.0.2=py311h5082efb_2
setuptools=80.9.0=pyhff2d567_0
simple-websocket=1.1.0=pypi_0
six=1.17.0=pypi_0
sympy=1.14.0=pyh04b8f61_5
tbb=2021.13.0=h18a62a1_3
tk=8.6.13=h2c6b04d_2
tomli=2.2.1=pyhe01879c_2
typing_extensions=4.15.0=pyhcf101f3_0
tzdata=2025.2=pypi_0
ucrt=10.0.26100.0=h57928b3_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h818238b_31
vcomp14=14.44.35208=h818238b_31
werkzeug=3.1.3=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
wsproto=1.2.0=pypi_0
xz=5.8.1=h208afaa_2
xz-tools=5.8.1=h2466b09_2
yaml=0.2.5=h6a83c73_3
gunicorn
eventlet
